Refactored Teams Page
